ReInc 0.6: implemented ranks for shuriken

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/buffs/PinCushion.java
@@ -71,7 +71,8 @@ public class PinCushion extends Buff {
 	@Override
 	public void detach() {
 		for (Item item : items)
-			Dungeon.level.drop( item, target.pos).sprite.drop();
+            if (item.quantity() > 0)
+			    Dungeon.level.drop( item, target.pos).sprite.drop();
 		super.detach();
 	}
